Modularize team crest routes in separate file
